Simple FullCallendar-style service

+++ src/main/java/com/planificateur/eventplanner/repository/LocalRepository.java
@@ -3,5 +3,5 @@ package com.planificateur.eventplanner.repository;
 import com.planificateur.eventplanner.models.Local;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalRepository extends JpaRepository<Local, Integer> {
+public interface LocalRepository extends JpaRepository<Local, Long> {
 }
